Move Timer.h to src/Ice
