+++ cpp/src/DataStorm/msbuild/datastorm/datastorm.vcxproj.filters
@@ -212,9 +212,6 @@
     <ClInclude Include="..\..\NodeSessionManager.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\Timer.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <SliceCompile Include="..\..\..\..\..\slice\DataStorm\SampleEvent.ice">
