+++ cpp/src/IceGrid/IceGridNode.cpp
@@ -2,12 +2,12 @@
 
 #include "../Ice/ConsoleUtil.h"
 #include "../Ice/FileUtil.h"
+#include "../Ice/Timer.h"
 #include "Activator.h"
 #include "DescriptorParser.h"
 #include "FileUserAccountMapperI.h"
 #include "Ice/Ice.h"
 #include "Ice/StringUtil.h"
-#include "Ice/Timer.h"
 #include "NodeAdminRouter.h"
 #include "NodeI.h"
 #include "NodeSessionManager.h"
